Update vendor list
Updated vendors from a vid file that was kindly provided to me from ODVA.
I wrote a small python script to read the ODVA file and output in a predefined
format, so i can update vendors in different tools.
To preserve obsolete vendors, i issue a list of them separately.